# src/common/documentation/api-documentation.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"access_token":"","refresh_token":""}}}}},"403":{"description":"","content":{"application/json":{"schema":{"example":{"code":403,"type":"FORBIDDEN","message":"Invalid credentials, please check inputs"}}}}}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"code":200,"type":"OK","message":"You have successfully log out"}}}}},"403":{"description":"","content":{"application/json":{"schema":{"example":{"code":403,"type":"FORBIDDEN","message":"Invalid credentials, please check inputs"}}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"access_token":"","refresh_token":""}}}}},"403":{"description":"","content":{"application/json":{"schema":{"example":{"code":403,"type":"FORBIDDEN","message":"You don't have permission to access this page"}}}}}},"tags":["auth"]}},"/auth/registration":{"post":{"operationId":"AuthController_registration","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"code":200,"type":"OK","message":"You have successfully registered"}}}}},"400":{"description":"","content":{"application/json":{"schema":{"example":{"code":400,"type":"BAD_REQUEST","message":"User with this email already exists"}}}}}},"tags":["auth"]}},"/user/active":{"get":{"operationId":"UserController_active","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"first_name":"","last_name":"","email":""}}}}},"400":{"description":"","content":{"application/json":{"schema":{"example":{"code":400,"type":"BAD_REQUEST","message":"Invalid user"}}}}}},"tags":["user"]}}},"info":{"title":"API documentation","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AuthDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RegistrationDto":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["first_name","last_name","email","password"]}}}}